requirements.txt
streamlit>=1.12.0
pandas>=1.3.0
numpy>=1.20.0
altair==4.2.2
plotly>=5.0.0
pyyaml>=6.0
asyncio
psycopg2-binary>=2.9.0